+++ cmd/gotraceui/canvas.go
@@ -67,8 +67,7 @@ const (
 
 	minSpanWidthDp unit.Dp = spanBorderWidthDp*2 + 4
 
-	spanBorderWidthDp            unit.Dp = 1
-	spanHighlightedBorderWidthDp unit.Dp = 2
+	spanBorderWidthDp unit.Dp = 1
 )
 
 const animateLength = 250 * time.Millisecond